fix ARM+msys2 size_t print

+++ tests/tmp_store.c
@@ -29,7 +29,7 @@ static OQS_STATUS oqs_fload(const char *fname, uint8_t *data, size_t len) {
 		return OQS_ERROR;
 	}
 	if (fread(data, len, 1, fp) != 1) {
-		fprintf(stderr, "Couldn't read all %ld bytes correctly (operations called in proper sequence?). Exiting.\n", len);
+		fprintf(stderr, "Couldn't read all %d bytes correctly (operations called in proper sequence?). Exiting.\n", (int)len);
 		return OQS_ERROR;
 	}
 	fclose(fp);
